requirements.txt
python-dotenv
langchain-core
langchain-openai
pydantic
python-pptx
pypdf
Pillow
PyMuPDF
pypdfium2 
comtypes 
pandas
openpyxl
python-dotenv
fastapi
uvicorn
python-multipart
qdrant-client
sentence-transformers